# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.11.0",
    "title": "SiSMoChat API",
    "description": "API for SiSMoChat - A secure chat application focused on child safety with parental control."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/connections/": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get server status message",
        "description": "Returns an optional message for client display at startup (e.g. maintenance notices).",
        "responses": {
          "200": {
            "description": "Status message (null if none)"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/connections/sent/{parent}": {
      "get": {
        "tags": [
          "Connections"
        ],
        "summary": "Get sent connection requests with status",
        "description": "Returns all connection requests initiated by children of this parent, with current status.",
        "parameters": [
          {
            "name": "parent",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sent requests with status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/connections/approvalList/{parent}": {
      "get": {
        "tags": [
          "Connections"
        ],
        "summary": "Get pending connection requests for approval",
        "description": "",
        "parameters": [
          {
            "name": "parent",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending connection requests"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/connections/{user}": {
      "get": {
        "tags": [
          "Connections"
        ],
        "summary": "Get connections list for a user (by parent)",
        "description": "",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max results (default 20)",
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Skip N results (default 0)",
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "List of connected user IDs"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/connections/{from}/{to}": {
      "post": {
        "tags": [
          "Connections"
        ],
        "summary": "Request a connection between two users",
        "description": "Parent requests a friendship connection between two child users.",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Connection requested"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Users not found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/connections/{connid}": {
      "patch": {
        "tags": [
          "Connections"
        ],
        "summary": "Accept or reject a connection request",
        "description": "Update connection status: 0=accepted, 2=rejected.",
        "parameters": [
          {
            "name": "connid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ConnectionStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Status updated"
          },
          "400": {
            "description": "Invalid status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Connection not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Connections"
        ],
        "summary": "Remove a connection (by parent)",
        "description": "Deletes both sides of a connection and notifies the other user.",
        "parameters": [
          {
            "name": "connid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Connection removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to remove this connection"
          },
          "404": {
            "description": "Connection not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ConnectionStatus": {
      "type": "object",
      "properties": {
        "status": {
          "type": "number",
          "example": 0
        }
      }
    }
  }
}